[Bugfix] Fix intermittent bug found in partial liveness checking.
Iff nextState itself is not on disk, its successor nodes
won't be either (or we don't know it's successors yet).
This scenario occurs when liveness checking is executed
on a *partial* graph only. The model checker has added
curState with nextState as its successor, nextState has
not been added yet though. Basically this is where the
partial graph has its border to the unknown.

Continuing the for loop is fine. But we might as well add
nextState to the queue. Once it gets picked up by the outer
while loop, it will be skipped anyway as its succCnt is 0.