[Bugfix] Stop NoSuchElementException during distributed modelchecking.
TLCIterator#getLast does not return the last element when the largest
fingerprint that is *not* on disk fails to be in the last non-null
bucket. This case however, is perfectly valid. Its likelyhood is
increased when checkpointing is enabled. Then the buckets are flushed in
shorter intervals causing a higher chance that the last bucket has only
on-disk fingerprints in it when getLast() is called.