[Bugfix] Trace exploration fails silently after loading an external TLC output file into the Model editor's result page.
Loading an existing TLC output file into the Model editor's results page
only creates the MC.out, but not the MC_TE.out file. Then trace
exploration fails silently due to the missing MC_TE.out file. It is
unknown why there has to be a MC_TE.out file which is an identical copy
of the MC.out file in the first place.