[Feature] IStateFunctor as parameter for ModelChecker#getInitStates(..)
Pass a functor to getInitStates() to - instead of adding all init states
into an intermediate StateVec to check and add each state in a
subsequent loop - directly check each state one-by-one and add it to the
queue, fingerprint set and trace file. This avoids allocating memory for
StateVec (which depending on the number of init states can grow to be
GBs) and the subsequent loop over StateVec.