[Feature] Ignore duplicate states (in terms of equality) when checking liveness.
This significantly improves the performance of safety checking with
liveness checking enabled because less redundant works has to be carried
out during insertion of states into the liveness graph. Namely, only
distinct states are evaluated against the liveness' consistency
properties (a usually expensive operation).
An additional advantage is, that the number of GraphNode transitions
is approximately known in advance. Thus, a large enough amount of
memory can be allocated in a single step.

A side effect is, that TLC now finds different counter-examples. This
is due to a changed order in which states are inserted into the
liveness graph. Where it previously inserted states in the order
returned by the next state generation, it now inserts the states in
the order of their fingerprints. However, the insertion order is
just a technical detail of TLC and has no relevance the TLC's
correctness.