[Tests] Liveness/Symmetry tests exhibiting bugs in TLC.
MC: It produces an error report on a correct property - or at least, one
on which TLC finds no error without symmetry.

MCa: It produces a bogus trace for an incorrect property. The trace is
bogus because the loop has enq[e1] = v2, which implies that ENABLED
EndEnq(e1) is continuously enabled but never executed - contradicting
the Liveness hypothesis.