Eagerly allocate GapTextStore's internal buffer according to log file size.
Document targets source code documents of a few thousand lines.
It quickly becomes inefficient for large inputs like e.g. a TLC log file
can be.

LargeTextStoreDocument thus drastically increases the min and max
sizes of the Document's underlying ITextStore.
Experiments revealed that this optimization significantly increases the
TagBasedTLCOutputIncrementalParser performance.

As we know the size of the log file a-prior, the ITextStore's
internal storage can be sized up-front to match the input's size.

[Bug][Toolbox][Performance]