Register TLCServer with two different names in RMI registry. One as a lookup mechanism for FPSets, the other for TLCWorkers.
This separation of lookups makes it possible to defer worker registration upon after the generation of init states, which however requires the presence of FPSets.

Registering workers during init state generation, has the effect that the breadth-first search extremely degrades when a (large number) of workers concurrently explore the StateQueue while the sequential master generates the potentially huge set of init states.

Other side effects are:
- A subset of workers can miss the termination/exit signal sent by the master when it find a violation among the init states and terminates
- It is convenient to be able to quickly fix the model upon a violating init state without being required to manually restart all workers

[Bug][TLC]