Switch from contained based travis builds to classic ones.
Classic builds have root access with which we disable
the OOM killer and kill unused processes to make the
memory-expensive JUnit tests pass.

The UI tests are too instable on travis which is why
we call ant directly. That way, the build is much
faster. The maven/tycho based build - which executes
UI tests - is commented.

[Build]