A variant of the spec which overlaps insertions and eviction. However, this variant violates Inv (all fps in history are either in primary or secondary). Suppose Thread A evicts fps to newsecondary and later eventually replaces secondary with newsecondary. During the time frame where a fingerprint is evicted to newsecondary, it cannot be marked as being on secondary. Another thread B might otherwise concurrently replace the fp marked as being on secondary with some fingerprint fp'. Now no thread will see fp which is only stored in newsecondary.
Changing contains to also check newsecondary, such that:

contains(f,t,seq,Q) ==
  \/ \E i \in 0..Q: isMatch(f,idx(f,i),t)
  \/ \E i \in 1..Len(seq): seq[i] = f
  \/ \E i \in 1..Len(newsecondary): newsecondary[i] = f

obviously removes the violation of Inv. However, an implementation
would suddenly have to check two disk files where one (newsecondary)
is being written to. That seems rather difficult to get right.

Another alternative appears to be to queue up marking evicted
fingerprints. The downside however is, that the queue can become
quite large up to the size of the set.

Making the other threads queue up insertions will result in a smaller
queue, but they can detect if eviction is ongoing or not.

For the moment, making insertion threads wait during eviction seems
to be both simple and sufficient. Insertions will most likely fail
anyway during eviction because of a high load factor. Additionally,
exclusive eviction makes it dead simple to minimize P to the smallest
possible value.
However, one can argue that the algorithm is no longer lock-free. It
is from the point of insertion threads for as long as there is no
eviction going on though. Also, we won't need CASN. Basic CAS will
do which is directly supported by Java.

[Documentation][TLC]