An internal error occurred during: "TLC run for Model_1". tlc2.tool.fp.OffHeapDiskFPSet.isSupported()Z
Fixes Github issue #133
https://github.com/tlaplus/tlaplus/issues/133

The underlying bug is a broken Toolbox update such that the Toolbox
continues to load an old version of the TLC bundle
(org.lamport.tlatools) instead of the newest one included in the update.
Adding "org.eclipse.update.configurator" to the configurations causes
the generated config.ini to contain org.eclipse.update.reconcile=false.
This statement - and simpleconfigurator set to startLevel=1 - asserts
that the Toolbox generally loads the most recent bundle versions.

[Bug][Toolbox]