Performance regression due to bugfix for Bugzilla #279.
Calling SetEnumValue#convert causes e.g. an instance of type SubsetValue
to be explicitly enumerated. Prior to the fix for #279, TLC was smart
enough to not enumerate all elements a-priori but to generate elements
one-by-one and return the first for which the given predicate is true.

See commit ec20a9b6217f25547d8a0f755a991394e72bb773 and
ffb267cf223337961e92f85c555f5825bd5705d2 on which this commit depends.

[Bug][TLC]