Bidirectional transitions cause TLC to incorrectly report the violation of a liveness property and print a bogus counterexample.
For efficiency reasons the intermediate representation of an SCC is just
a set of nodes without transitions.  It is thus not an actual graph!

A: The pair (Spec, ~Property) is split into three conjuncts: <>[]A,
[]<>A and []<>S  but lets ignore []<>S in this context because it is not
relevant.   When TLC runs Tarjan's algorithm (TA), it skips transitions
that do not satisfy <>[]A (coined "EA actions" in TLC) and consequently
does not find some SCCs.  Lets call the set of skipped SCCs IGN (for
ignored).  The SCCs however that TA does find are further checked for
[]<>A.  Let CHK be the set of such SCCs.

B: To check []<>A on C \in CHK, TLC non-deterministically picks a node
from C and explores all the node's outgoing transitions that lead to
another node in C (transitions leading to nodes not in C lead out of the
SCC).  For each transition it does explore, TLC checks []<>A . If the
transition satisfies []<>A TLC has found a valid counter-example.


Given an SCC is just a set of nodes, lets now suppose CHK \intersect IGN
# {}.  In other words, there exists an SCC \in CHK and \in IGN (this is
e.g. the case for the two specs discussed in the GitHub issue).  Due to
the algorithm above, even transitions "belonging" to the SCC \in IGN are
[]<>A checked.  If any one of these transitions satisfies []<>A, TLC
incorrectly reports ~Property to be violated.

The fix changes TLC to not only check []<>A in B: but also
re-check<>[]A.  Only iff both are satisfied, the SCC violates ~Property.

A similar bug surfaces when TLC correctly finds ~Property to
be violated.  Unless TLC re-checks <>[]A when it computes the minimal
SCC, it reports a bogus counter-example with C transitions.


Fixes GitHub issue #201
https://github.com/tlaplus/tlaplus/issues/201

[Bugfix][TLC][Changelog]