Display a warning when liveness is combined with symmetry.
TLC may fail to find (liveness) violations when liveness checking is
combined with symmetry reduction.  To warn users, the Toolbox shows a
warning message on the model editor when a property is defined in
combination with one or more symmetry sets.  If the properties happen to
be a safety properties only, the Toolbox erroneously shows the warning
(the Toolbox cannot discern liveness from safety properties).  To
reliably warn users, the Toolbox delays showing the warning until TLC
has started and printed a corresponding warning to the console.

The old, unreliable warning has been downgraded to an informational
message instructing users to pay attention to the warning at runtime.

[Refactor][Toolbox]