Usage data reporting (startup banner) for TLC.
Part I: The actual reporting iff a user opted-in to have usage data
reported.  In order to opt in, a ~/.tlaplus/udc.txt file has to exist
and its first line must not match the string "NO_UDC".

When a user has opted in, the first line of ~/.tlaplus/udc.txt is
truncated to a 32 character string and transmitted as part of the usage
data.  If the string equals "RANDOM_IDENTIFIER", TLC will choose a UUID
as the (non-user-identifiable) id.

An exemplary usage data record:

[18/Apr/2019:19:23:00 +0200] "HEAD
/?ver=development,mode=bfs,workers=1,cores=4,osName=Linux,osVersion=4.18.0-16-generic,osArch=amd64,jvmVendor=Oracle+Corporation,jvmVersion=11.0.2,jvmArch=x86_64,jvmHeapMem=7851,jvmOffHeapMem=64,seed=-4016958957504178886,fpidx=1,jvmPid=6230,fpset=MSBDiskFPSet,queue=DiskStateQueue,toolbox=false,id=5r423piogfolf89z983489t68q3c4foa,ts=1555608180252,optout=false
HTTP/1.1"

[Feature][TLC][Changelog]