Profiling shows states founds and distinct states multiple times for one action name.
Determine if the mapping from the action's name/identifier/declaration
to the action's definition is 1:1 or 1:N.

Act == /\ x  = 23
       /\ x' = 42

vs:

Act == \/ /\ x  = 23
          /\ x' = 42
       \/ /\ x  = 123
          /\ x' = 4711

or:

Act == (x  = 23 /\ x' = 42) \/ (x  = 123 /\ x' = 4711)

For a 1:1 mapping this prints just the location of Act. For a 1:N
mapping it prints the location of Act _and_ the location (in shortened
form) of the actual disjunct.

[Bug][TLC]