Copy/Export error trace to clipboard.
Exports the error trace (with filtering and trace expressions applied)
to the system's clipboard formatted as TLA+ (sequence of records).  If
the shift key is pressed, the exported trace omits the "_TEAction"
record with fields "position", "name", and "location".

Exports the states selected in the treeviewer or all if no selection
given (disables export if variables are part of the tree's selection).

I'd have preferred to let users optionally provide a TLA+ expression
that is evaluated on the trace before it is exported.  That would have
allowed to re-format to e.g. JSon.  However, this would have required
more substantial changes of the trace explorer.  As a workaround, the
exported trace can be pasted into "Evaluate Constant Expression" and
reformatted (a REPL would be nice).

This commit provides:

- Workaround for GitHub issue #344 (Raw Trace-Expression error trace has
cryptic action names)
https://github.com/tlaplus/tlaplus/issues/344
- A part of the puzzle for the latest design discussed as part of GitHub
issue #191 (Animation Support for Traces)
https://github.com/tlaplus/tlaplus/issues/191

[Feature][Toolbox][Changelog]