When model-checking specifications with TLC, it is occasionally useful to leave the mathematical beauty of TLA (behaviors are not part of the universe of discourse. Formulas such as \E state \in Behavior: state... cannot be stated) and evaluate expressions on (a prefix of) a behavior (or sets of behaviors if collected with TLCGet/TLCSet). One particular example is post-processing of a TLC error trace to "pipe" the trace to ShiViz or Will's TLA+ animator.  Another example is to translate an error trace from its TLA+ representation to Json (see Json.tla in CommunityModules), which when done in 3rd party scripts, has proven to be brittle (please don't write ad-hoc TLA+ parsers!!!).
Let SG be the state graph explored by TLC and let FG be a fingerprint
(hash) graph that is isomorphic to SG (ignoring symmetry reduction).
Let fp(s) be the function that maps the state s of SG to the
corresponding fingerprint of FG.  ConcurrentTLCTrace#getTrace takes as
input a state s_n in SG and returns the sequence of states p_s on the
(shortest) path in the state graph from s_n to an initial state s_0.
For efficiency reasons though, TLC does not persist SG nor FG.  Instead,
TLC writes a trace T, which is the set of (shortest) paths (forest of
trees) from any fingerprint in FG back to a fingerprint of an initial
state.  Thus, TLC traverses the path in T from fp(s_n) to fp(s_) and
obtains a sequence of fingerprints p_f that correspond to a path in FG.
To convert p_f into p_s, TLC re-runs ordinary state space exploration,
except that it ignores all states whose fingerprints are not on the path
p_f (it follows the sequence of fingerprints).  The number of states
generated might still be large depending on the number of successor
states.

More concretely, T is a set of <<address, fingerprint>> tuples.
In the case of TLCTrace, the address is just the offset of a
fingerprint's predecessor in the file that persists T.  The offset is
simply called "ptr".  To construct the fingerprint path (p_f) from
above, TLC iteratively reads the predecessor fingerprint of s_n by
traversing the sequence of addresses.  Note that TLC assigns each new
state the address of its predecessor during state space exploration.
In the case of ConcurrentTLC, the address is a pair <<offset, workerId>>
where workerId identifies a TLC worker. Let N be the number of TLC
workers, the concurrent trace partitions T into N fragments s.t. the
union of the fragments is T (which is due to minimize (lock) contention
when writing T). Thus, to construct p_f, the address of a predecessor
fingerprint is determined by its ptr *and* its workerId.

Previously, TLC reconstructed p_s s.t. the address for all states on p_s
was undefined (null), which was fine because the path was only ever
printed as the error trace.  Today, however, the `Trace` and `TraceFrom`
operators from the CommunityModules treat p_s as a first-class object
(see above).   It thus might be that a reconstructed state of p_s is
input to ConcurrentTLCTrace#getTrace, which would fail when its address
is undefined.  Below is an example spec where this is the case:

---- MODULE Example ----
EXTENDS Integers, TLCExt
VARIABLE x, y

Init ==
    x = 0 /\ y = 1

Next ==
    /\ x' = Trace[y].x + 1
    /\ y' = y + 1
=====

Note that this is highly inefficient because TLC will (recursively)
reconstruct the trace of each state on p_s, which -in the best case- is
O(n^2).  I expect `Trace` and `TraceFrom` primarily to appear in
invariants and TLC state constraints; where it doesn't have to be
evaluated recursively.

[Bug][TLC]