Fix race condition in liveness counterexample reconstruction:
```
Error: TLC threw an unexpected exception.
This was probably caused by an error in the spec or model.
See the User Output or TLC Console for clues to what happened.
The exception was a java.io.IOException
: Read past end of file (increase with setLength)?
```

This commit fixes a bug introduced in TLC 2.08 (2015) when the
reconstruction of lasso-shaped liveness counterexamples was parallelized
(construct the prefix leading to the lasso state in parallel with the
lasso itself).  The oversight was that the underlying
AbstractDiskGraph#getNode & AbstractDiskGraph#getNodeFromDisk methods
are not thread-safe.  Although there are no disk *writes* during the
construction of a counterexample, the file pointers -internal to
BufferedRandomAccessFile- can be corrupted by concurrent reads.  The
method-level synchronization in AbstractDiskGraph#getNodeFromDisk
addresses this issue. The synchronization in AbstractDiskGraph#getNode
takes care of a more pathological race condition where the GraphNode
cache (AbstractDiskGraph#gnodes) could be corrupted if two states hash
to the same array position.

Fixes Github issue #597
https://github.com/tlaplus/tlaplus/issues/597

[Bug][TLC][Regression][Changelog]