Combine original Spec with generated action constraint instead of generating artificial _init and _next.
This approach has the following advantages:
* Original action names in the error-trace
* Users can change the original spec and quickly confirm that the change
addresses the error-trace (rerunning the trace spec stops generating
states).

The following problems with this approach became apparent:
* The approach is incompatible with specs that involve randomization
(the order in which states are generated differs from the original TLC
run). This might be an acceptable restriction.
* TLC generates many more states, which is especially a problem with
specs with huge sets of initial states.

A workaround that preserves action names of successor steps while
minimizing to a single initial state, would be to re-define only Init in
the trace spec.  However, not all specs (explicitly) define a Init that
can be re-defined.

[Feature][TLC][WIP][Abandoned]