Let TLA+ debugger automatically halt (unless nohalt) when an invariant is violated in simulation mode.
Technically, this feature uses the debugger's infrastructure to handle
EvalExceptions and TLCRuntimeExceptions.  In addition, the debugger
now handles InvariantViolationExceptions.  Unfortunately, this restricts
this feature to simulation mode, which first throws IVE before it prints
the error-trace (potentially evaluating an alias).  In model-checking
mode, TLC first prints the error-trace and only then throws an IVE.

[Feature][TLC]