TLA+ debugger support for:
In addition to evaluation exceptions, the debuggers now halts on
violated invariants (box properties) and missing unchanged statements
unless halting is turned off explicitly.  The debugger handles all
exceptions only once at the call site.

Reset (stepping back) to the previous stack-frame and initial
stack-frame, where a frame is defined as a point in the evaluation, from
which the
the interpreter (Tool) can resume.  However, resetting to an earlier
frame interferes with BFS because TLC's queue & seen set do *not* get
reset.

[Feature][TLC][Debugger]