Exclude a stack-frame's "main" state in the "Trace" of the variable view.
The TLCStateStackFrame shows the main state in the "State" node, the
action frame shows its main state (the successor of s -> t) in the
"Action" node, and the successor frame shows 0..N states under the
"Successors" node.  In other words, the "Trace" node shows the states
leading up to the main state.

[Feature][TLC][Debugger]